requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pyyaml>=5.4.0
matplotlib>=3.4.0

# Optional dependencies for advanced features
# scikit-optimize>=0.9.0  # For Bayesian optimization in LQR tuning
# pytest>=6.0.0  # For running tests
